+++ crates/cargo-contract/src/cmd/mod.rs
@@ -21,6 +21,7 @@ pub mod encode;
 pub mod info;
 pub mod instantiate;
 pub mod remove;
+pub mod schema;
 pub mod storage;
 pub mod upload;
 pub mod verify;
@@ -38,6 +39,10 @@ pub(crate) use self::{
     },
     instantiate::InstantiateCommand,
     remove::RemoveCommand,
+    schema::{
+        GenerateSchemaCommand,
+        VerifySchemaCommand,
+    },
     storage::StorageCommand,
     upload::UploadCommand,
     verify::VerifyCommand,
@@ -123,7 +128,7 @@ impl CLIExtrinsicOpts {
     }
 }
 
-const STORAGE_DEPOSIT_KEY: &str = "Storage Deposit";
+const STORAGE_DEPOSIT_KEY: &str = "Storage Total Deposit";
 pub const MAX_KEY_COL_WIDTH: usize = STORAGE_DEPOSIT_KEY.len() + 1;
 
 /// Print to stdout the fields of the result of a `instantiate` or `call` dry-run via RPC.
@@ -232,8 +237,13 @@ pub fn basic_display_format_extended_contract_info(info: &ExtendedContractInfo) 
         MAX_KEY_COL_WIDTH
     );
     name_value_println!(
-        "Storage Deposit",
-        format!("{:?}", info.storage_item_deposit),
+        "Storage Items Deposit",
+        format!("{:?}", info.storage_items_deposit),
+        MAX_KEY_COL_WIDTH
+    );
+    name_value_println!(
+        STORAGE_DEPOSIT_KEY,
+        format!("{:?}", info.storage_total_deposit),
         MAX_KEY_COL_WIDTH
     );
     name_value_println!(